requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
motor==2.5.1
pymongo==3.12.0
pydantic[email]==1.8.2
python-dotenv==0.19.0